pyproject.toml
[project]
name = "ma-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.42.0",
    "fastapi[standard]>=0.115.6",
    "openai>=1.58.1",
    "phidata>=2.7.6",
    "psycopg>=3.2.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "supabase>=2.11.0",
    "twilio>=9.4.1",
]
